+++ src/heuristicvalue.h
@@ -1,0 +1,6 @@
+#ifndef HEURISTIC_VALUE_H
+#define HEURISTIC_VALUE_H
+
+using HeuristicValue = int;
+
+#endif // HEURISTIC_VALUE_H