+++ CHANGELOG.md
@@ -10,3 +10,21 @@ All notable changes to this project will be documented in this file.
 
 * Added main structure of this program just to cover Tetris game requirements
 * Added scripts that should update Tetris.pro and its config.h files to enable this code
+
+## 15.08.2024
+
+* Added heuristic function that can simulate each move and calculate score of that position
+
+## 18.08.2024
+
+* Fixed issue with calculating gaps
+
+## 19.08.2024
+
+* Fixed height drop calculation
+* Added bumpiness as heuristic measurement
+* Updated project to working version
+
+## 20.08.2024
+
+* Added holding shape checking and swapping ability